# server/docs/swagger.yaml
definitions:
  api.BatchDiscoverRequest:
    properties:
      category:
        type: string
      deepScan:
        type: boolean
      rootDomains:
        items:
          type: string
        type: array
    required:
    - rootDomains
    type: object
  api.BatchIdsRequest:
    properties:
      ids:
        items:
          type: integer
        type: array
    required:
    - ids
    type: object
  api.DiscoverSubdomainsRequest:
    properties:
      category:
        type: string
      deepScan:
        type: boolean
      rootDomain:
        type: string
    required:
    - rootDomain
    type: object
  api.IgnoreDomainRequest:
    properties:
      domainId:
        type: integer
      ignore:
        type: boolean
    required:
    - domainId
    type: object
...
  api.ReProbeRequest:
    properties:
      rootDomain:
        type: string
    required:
    - rootDomain
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.8.8
paths:
...
  /certAdvanced/batchDiscoverSubdomains:
    post:
      consumes:
      - application/json
      parameters:
      - description: 批量发现参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.BatchDiscoverRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 批量子域名发现成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量子域名发现
      tags:
      - CertAdvanced
  /certAdvanced/batchIgnore:
    post:
      consumes:
      - application/json
      parameters:
      - description: IDs
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.BatchIdsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 批量忽略成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量忽略
      tags:
      - CertAdvanced
  /certAdvanced/batchReprobe:
    post:
      consumes:
      - application/json
      parameters:
      - description: IDs
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.BatchIdsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 批量探测完成
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 批量重新探测
      tags:
      - CertAdvanced
  /certAdvanced/discoverSubdomains:
    post:
      consumes:
      - application/json
      parameters:
      - description: 发现参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.DiscoverSubdomainsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 子域名发现成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 子域名发现
      tags:
      - CertAdvanced
  /certAdvanced/exportSubdomainReport:
    get:
      consumes:
      - application/json
      parameters:
      - description: 根域名
        in: query
        name: rootDomain
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: 文件流
          schema:
            type: file
      security:
      - ApiKeyAuth: []
      summary: 导出子域名报告
      tags:
      - CertAdvanced
  /certAdvanced/getDomainAssetList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - in: query
        name: category
        type: string
      - in: query
        name: certId
        type: integer
      - in: query
        name: certStatus
        type: integer
      - in: query
        name: certType
        type: string
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - in: query
        name: domain
        type: string
      - in: query
        name: domainStatus
        type: integer
      - in: query
        name: ip
        type: string
      - in: query
        name: isIgnored
        type: boolean
      - description: 关键字
        in: query
        name: keyword
        type: string
      - in: query
        name: lastProbedAt
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: parentId
        type: integer
      - in: query
        name: rootDomain
        type: string
      - in: query
        name: source
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取域名资产列表
      tags:
      - CertAdvanced
  /certAdvanced/getDomainTree:
    get:
      consumes:
      - application/json
      parameters:
      - description: 根域名
        in: query
        name: rootDomain
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  additionalProperties: true
                  type: object
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取域名树
      tags:
      - CertAdvanced
  /certAdvanced/ignoreDomain:
    post:
      consumes:
      - application/json
      parameters:
      - description: 忽略参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.IgnoreDomainRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 忽略/取消忽略域名
      tags:
      - CertAdvanced
  /certAdvanced/reProbeDomainTree:
    post:
      consumes:
      - application/json
      parameters:
      - description: 探测参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/api.ReProbeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 重新探测成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 重新探测域名树
      tags:
      - CertAdvanced
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"
...